Add {de,}serialization implementations.
The following types can now be both serialized and deserialized.
* Selector
* ColorSetting
* State

Zoned and Random<T> can both be serialized but not deserialized, as
the only real use case for deserialization for these is if users want
to do something like "all:random".parse::<Random<Selector>>().

std::time::Duration also has custom {de,}serialization for compatibility
with the LIFX API (serializes as a float with millisecond precision for
now).

This closes #2.